requirements.txt
absl-py==2.1.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
cloudpickle==3.0.0
cmake==3.28.4
contourpy==1.2.0
cycler==0.12.1
Cython==0.29.36
einops==0.7.0
etils==1.5.2
evdev==1.7.0
fasteners==0.19
filelock==3.9.0
fonttools==4.50.0
fsspec==2023.4.0
glfw==2.7.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.10.0
idna==3.4
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.2
kiwisolver==1.4.5
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
mujoco==3.1.3
mujoco-py==2.1.2.14
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.1
numpy==1.24.1
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
psutil==5.9.8
pycparser==2.21
Pygments==2.17.2
pynput==1.7.6
PyOpenGL==3.1.7
pyparsing==3.1.2
pyrallis==0.3.1
python-dateutil==2.9.0.post0
python-xlib==0.33
PyYAML==6.0.1
requests==2.28.1
rich==13.7.1
scipy==1.12.0
six==1.16.0
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.26.13
zipp==3.18.1
wandb==0.16.4
